Add db file path support
